tests/wpt/web-platform-tests/css/css-scroll-snap/parsing/scroll-snap-type-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap: getComputedStyle().scrollSnapType</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-snap-type">
<meta name="assert" content="scroll-snap-type computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("scroll-snap-type", "none");

test_computed_value("scroll-snap-type", "x");
test_computed_value("scroll-snap-type", "y");
test_computed_value("scroll-snap-type", "block");
test_computed_value("scroll-snap-type", "inline");
test_computed_value("scroll-snap-type", "both");

test_computed_value("scroll-snap-type", "y mandatory");
test_computed_value("scroll-snap-type", "inline proximity", "inline");
</script>
</body>
</html>
